+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 - **Breaking Change**: not reset exit status before call user-defined command due to posix shell compatibility
 - **Breaking Change**: not allow ``/`` character in user-defined command name
 - **Breaking Change**: change cli error message. now print command name
+- **Breaking Change**: now max input size is 1GB
 - auto-unwrap option type expression in for-in expression
   ```
   for a in "123" as String? { echo $a; }
@@ -98,6 +99,7 @@
 - improve internal error checking of ``FD`` iteration
     - now report io error
     - correctly check read string size limit
+- ``Module#_func`` method can accept up to 1GB input string
 
 #### Module